GUI finished

+++ pool.sh
@@ -24,31 +24,47 @@ chdir = /
 EOT
 }
 
-# add FPM pool
-_pool_add() {
-    declare pooldef path="$PHP_PATH/$PHPV/fpm/pool.d"
+# check site pool
+# $1 - name ($NAME)
+_pool_check() {
+    declare phpv name="${1:-$NAME}"
 
-    # PHP version isn't installed
-    [[ ! -d $path ]] && addmsg "PHP version '$PHPV' isn't installed or is awkward." $MSG_TYPE_ERR
-    # pool definition already exists
-    [[ -f $path/$NAME.conf ]] && addmsg "Pool '$NAME' PHP $PHPV FPM definition already exists." $MSG_TYPE_ERR
-
-    ((ERR_CNT > 0)) && return 1
-
-    pooldef="$(pool_tpl "$NAME" "$SITE_USER" "$SITE_GROUP" "$LISTEN_OWNER" "$LISTEN_GROUP")"
-    write "$pooldef" "$path/$NAME.conf" &&
-        addmsg "Pool '$NAME' PHP $PHPV FPM definition created."
-    return 0
+    for phpv in "${PHP_LIST[@]}"; do
+        if [[ -f $PHP_PATH/$phpv/fpm/pool.d/$name$CFG_EXT ]]; then
+            echo "$phpv"
+            return 0
+        fi
+    done
+    return 1
 }
 
 # remove FPM pool
 _pool_rm() {
     declare phpv path
 
-    for phpv in "${PHP_LIST[@]}"; do
-        path="$PHP_PATH/$phpv/fpm/pool.d/$NAME.conf"
-        [[ -f $path ]] && sudo rm "$path" && addmsg "Pool '$NAME' PHP $phpv FPM definition removed."
-    done
+    phpv="$(_pool_check)"
+    [[ -z $phpv ]] && return 1
+    path="$PHP_PATH/$phpv/fpm/pool.d/$NAME$CFG_EXT"
+    [[ -f $path ]] && sudo rm "$path" && addmsg "Pool '$NAME' PHP $phpv FPM definition removed."
+}
+
+# add FPM pool
+_pool_add() {
+    declare pooldef phpv path="$PHP_PATH/$PHPV/fpm/pool.d"
+
+    # PHP version isn't installed
+    [[ ! -d $path ]] && addmsg "PHP version '$PHPV' isn't installed or is awkward." $MSG_TYPE_ERR
+    # pool definition already exists
+    [[ -f $path/$NAME$CFG_EXT ]] && addmsg "Pool '$NAME' PHP $PHPV FPM definition already exists." $MSG_TYPE_ERR
+
+    ((ERR_CNT > 0)) && return 1
+
+    _pool_rm
+
+    pooldef="$(pool_tpl "$NAME" "$SITE_USER" "$SITE_GROUP" "$LISTEN_OWNER" "$LISTEN_GROUP")"
+    write "$pooldef" "$path/$NAME$CFG_EXT" &&
+        addmsg "Pool '$NAME' PHP $PHPV FPM definition created."
+    return 0
 }
 
 # FPM pools management
@@ -57,12 +73,10 @@ pool() {
     LONG=force,name:,php:
     _optarg "$@"
     msgclr
-    if [[ $CMD == $CMD_ADD ]]; then
-        _pool_add
-    elif [[ $CMD == $CMD_RM ]]; then
-        _pool_rm
-    else
-        addmsg "Command not recognized: $CMD" $MSG_TYPE_ERR
-    fi
+    case $CMD in
+        $CMD_ADD) _pool_add ;;
+        $CMD_RM) _pool_rm ;;
+        *) addmsg "Command not recognized: $CMD" $MSG_TYPE_ERR ;;
+    esac
     msgout
 }
